Refactoring
The main.cpp was around 1000 lines (including the old code commented
out), so I felt it was time to finally split everything into proper
files. It compiles and runs and there should be no change in behaviour.

Parser.h + ~.cpp:
    The Parser class and all its functionality

arch.h:
    Architecture related things, like the instruction set

debug.h:
    error(), log(), verbose(), fatal(), nameof() macros

utility.h:
    Free functions for doing string operations, checking whether instrs
    expect args or not, etc.
